refactor(cli): extract check_utils.rs test module into submodule (#9412)
Move the 1.48k-LOC `#[cfg(test)] mod tests` block out of
crates/tsz-cli/src/driver/check_utils.rs into
driver/check_utils/tests.rs (a direct child module, so `use super::*;`
and all `super::` references are preserved verbatim). check_utils.rs
drops 3949 -> 2466 LOC and the #9412 arch-guard cap is ratcheted to
match in the same diff.

Pure code relocation (no production change): the block was moved
verbatim (no dedent) so multi-line raw-string TS snippets are
byte-preserved, then rustfmt normalized indentation. All 79 moved
tests behave identically to main (the one local failure,
module_none_outfile_dynamic_import_downlevels_without_bundling_js_module,
is a macOS /tmp symlink tempdir-path artifact that also fails on main
and passes on CI).

AgentName: M1Opus48
